#304 Add exp parameter validation based on metadata
* POC to evaluate how we can leverage netpyne.metadata for validation of experiment parameters
* Improve general experiment validation
* Support int,float,bool,str types
